requirements.txt
streamlit
python-dotenv
faiss-cpu
sentence-transformers
pymupdf
markdown
unstructured
unstructured[pdf]
unstructured[docx]
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
tavily-python
beautifulsoup4